This is an improvement of the  InternetSpeedTest_Reworked_Final.zip
The improvement is that there is a check added for slower connections, on the previous version only up to 200mbps could be checked, since the file gets downloaded fast. Also there is a debug version added for testing a large file for a long time in order to check the average download speed, the debug is only for PCs, the browser has to work for a long time, while the using the PC normally.