ESP8266 nodeMCU as a WiFI repeater or mesh
Use the web flash tool(supported by chrome)
https://esp.huhn.me/

Download the friles from:
https://github.com/martin-ger/esp_wifi_repeater/tree/master

Extract the files and open the firmware folder

Upload the files to the nodeMUC with https://esp.huhn.me/
0x00000.bin at 00000
0x02000.bin at 02000
0x82000.bin at 0x82000

Power off and power on the NodeMCU ESP8266
Connect to the MyApp WiFi
Visit 192.168.4.1

Enter the SSID to repat and PW. Use the mesh option on the top. Wait 2 min.
Check the serial terminal on https://esp.huhn.me/ to see if the mesh is successfull.