Rename KiCad_28. AutomotiveBlinker_v1.zip to 28. KiCad_AutomotiveBlinker_v1.zip
An automotive cool blinker. Also measures the battery volttage.

Power supply is LM2576 switching at 50KHz.
MCU is nodeMCU wemos D1 mini.
LEDs.

Pin9 is connected to A0 to measure the car batt.
Don't forget the nodeMCU internal divider on A0. Voltage supplied on A0 is divided.

code.ino is the nodeMCU code.

Project is not tested. There may be bugs.

Everythin and the files are in the video description.